#504 [feat] 예외 처리 및 로깅 관련 모듈 이동 및 로깅 info level 추가
