+++ module-api/src/main/java/com/mile/controller/user/facade/AuthFacade.java
@@ -17,7 +17,7 @@ import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Component;
 
 @Component
-@RequiredArgsConstructor(access = AccessLevel.PRIVATE)
+@RequiredArgsConstructor(access = AccessLevel.PROTECTED)
 public class AuthFacade {
 
     private final UserService userService;
