+++ module-api/src/main/java/com/mile/exception/log/discord/model/Author.java
@@ -1,12 +1,13 @@
-package com.mile.log.discord.model;
+package com.mile.exception.log.discord.model;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class Footer {
+public class Author {
 
-    private final String text;
+    private final String name;
+    private final String url;
     private final String iconUrl;
 }
