+++ module-api/src/main/java/com/mile/exception/log/discord/model/Image.java
@@ -1,11 +1,12 @@
-package com.mile.log.discord.model;
+package com.mile.exception.log.discord.model;
+
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class Thumbnail {
+public class Image {
 
     private final String url;
 }
