+++ module-api/src/main/java/com/mile/exception/log/discord/model/Thumbnail.java
@@ -1,12 +1,11 @@
-package com.mile.log.discord.model;
-
+package com.mile.exception.log.discord.model;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public class Image {
+public class Thumbnail {
 
     private final String url;
 }
