+++ module-api/src/main/java/com/mile/exception/log/filter/FilterConfig.java
@@ -1,4 +1,4 @@
-package com.mile.log.filter;
+package com.mile.exception.log.filter;
 
 
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
