+++ module-api/src/main/java/com/mile/exception/log/filter/wrapper/CachedBodyServletInputStream.java
@@ -1,4 +1,4 @@
-package com.mile.log.filter.wrapper;
+package com.mile.exception.log.filter.wrapper;
 
 import jakarta.servlet.ReadListener;
 import jakarta.servlet.ServletInputStream;
