+++ module-api/src/main/java/com/mile/exception/log/utils/HttpRequestUtil.java
@@ -1,7 +1,7 @@
-package com.mile.log.utils;
+package com.mile.exception.log.utils;
 
 
-import com.mile.log.filter.wrapper.CachedBodyRequestWrapper;
+import com.mile.exception.log.filter.wrapper.CachedBodyRequestWrapper;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import lombok.AccessLevel;
