+++ module-api/src/main/java/com/mile/exception/log/utils/StringUtil.java
@@ -1,4 +1,4 @@
-package com.mile.log.utils;
+package com.mile.exception.log.utils;
 
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
