+++ module-common/build.gradle
@@ -4,6 +4,4 @@ jar { enabled = true }
 dependencies {
     //Aop
     implementation 'org.springframework.boot:spring-boot-starter-aop'
-    //Sentry
-    implementation 'io.sentry:sentry-spring-boot-starter-jakarta:7.9.0'
 }
